+++ web/src/beta/features/Editor/Map/LayerStylePanel/Editor/NodeSystem/Marker/Nodes/Height.tsx
@@ -1,28 +1,30 @@
 import { NumberInput } from "@reearth/beta/lib/reearth-ui";
-import { PolylineAppearance } from "@reearth/core";
+import { MarkerAppearance } from "@reearth/core";
 import { useT } from "@reearth/services/i18n";
 import { useNotification } from "@reearth/services/state";
 import { FC, useEffect, useState } from "react";
 
-import { LayerStyleProps } from "../../InterfaceTab";
-import NodeSystem from "../../NodeSystem";
-import ConditionalTab from "../../NodeSystem/ConditionTab";
-import ExpressionTab from "../../NodeSystem/ExpressionTab";
+import NodeSystem from "../..";
+import { LayerStyleProps } from "../../../InterfaceTab";
+import ConditionalTab from "../../ConditionalTab";
+import ExpressionTab from "../../ExpressionTab";
 
-const StrokeWidthNode: FC<LayerStyleProps> = ({
+const HeightNode: FC<LayerStyleProps> = ({
   optionsMenu,
   layerStyle,
   setLayerStyle
 }) => {
-  const [value, setValue] = useState<PolylineAppearance["strokeWidth"]>(
-    layerStyle?.value.polyline?.strokeWidth ?? 0
+  const [value, setValue] = useState<MarkerAppearance["height"]>(
+    (layerStyle?.value.marker?.height as number) ?? 0
   );
+
   const t = useT();
   const [, setNotification] = useNotification();
 
   useEffect(() => {
-    if (layerStyle?.value.polyline?.strokeWidth)
-      setValue(layerStyle?.value.polyline?.strokeWidth);
+    if (layerStyle?.value.marker?.height) {
+      setValue(layerStyle.value.marker.height);
+    }
   }, [layerStyle]);
 
   useEffect(() => {
@@ -33,9 +35,9 @@ const StrokeWidthNode: FC<LayerStyleProps> = ({
           ...prev,
           value: {
             ...prev.value,
-            polyline: {
-              ...prev.value?.polyline,
-              strokeWidth: value
+            marker: {
+              ...prev.value.marker,
+              height: value
             }
           }
         };
@@ -43,7 +45,7 @@ const StrokeWidthNode: FC<LayerStyleProps> = ({
     } catch (_e) {
       setNotification({ type: "error", text: t("Invalid style") });
     }
-  }, [setLayerStyle, setNotification, setValue, t, value]);
+  }, [value, setLayerStyle, setNotification, t]);
 
   const renderContent: Record<string, JSX.Element> = {
     value: <NumberInput value={value} onChange={setValue} />,
@@ -54,11 +56,12 @@ const StrokeWidthNode: FC<LayerStyleProps> = ({
       </ConditionalTab>
     )
   };
+
   return (
-    <NodeSystem title="StrokeWidth" optionsMenu={optionsMenu}>
+    <NodeSystem title="Height" optionsMenu={optionsMenu}>
       {(activeTab) => renderContent[activeTab] || null}
     </NodeSystem>
   );
 };
 
-export default StrokeWidthNode;
+export default HeightNode;
