+++ web/src/beta/features/Editor/Map/LayerStylePanel/Editor/NodeSystem/Marker/Nodes/index.tsx
@@ -1,7 +1,7 @@
 import { FC } from "react";
 
-import { LayerStyleProps } from "../../InterfaceTab";
-import HeightReferenceNode from "../../SharedNode/HeightReference";
+import { LayerStyleProps } from "../../../InterfaceTab";
+import HeightReferenceNode from "../../../SharedNode/HeightReference";
 
 import HeightNode from "./Height";
 import PointColorNode from "./PointColor";
